Loop created to iterate through positions and velocities
I created the loop to iterate finding the positions and velocities and store these in the list. I then plotted these against time, and noticed that the simulation gives a harmonic oscillation.
However, the limitations of this basic simulation was observed. If the sample size is not small enough, a damping effect is noticed, due to the method used to find the next position and velocity. This method might be suitable for the purpose of our project, but a slightly more sophisticated one might be better. We could use this to introduce the limitations of computational physics, and the tradeoff between run-time and precision.