Introduced kinetic energy function and slider
Using predefined functions, functions for the kinetic energy, potential energy and total mechanical energy were created.

A slider was also introduced to adjust the plots for initial velocity and mass of the SHO.

Encountered an issue where the kinetic energy is double the potential energy - predefined function may need revision.